Add visualization using trunk

+++ libs/simulation/src/lib.rs
@@ -12,9 +12,9 @@ pub struct World {
 
 #[derive(Debug)]
 pub struct Animal {
-    position: na::Point2<f32>,
-    rotation: na::Rotation2<f32>,
-    speed: f32,
+    position: na::Point2<f64>,
+    rotation: na::Rotation2<f64>,
+    speed: f64,
 }
 
 #[derive(Debug)]
@@ -74,7 +74,7 @@ impl Animal {
         }
     }
 
-    pub fn position(&self) -> na::Point2<f32> {
+    pub fn position(&self) -> na::Point2<f64> {
         // ------------------ ^
         // | No need to return a reference, because na::Point2 is Copy.
         // |
@@ -90,7 +90,7 @@ impl Animal {
         self.position
     }
 
-    pub fn rotation(&self) -> na::Rotation2<f32> {
+    pub fn rotation(&self) -> na::Rotation2<f64> {
         self.rotation
     }
 }
